# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LifeBridge AI",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ai/trust-check": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Trust Check",
        "operationId": "trust_check_api_v1_ai_trust_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScamCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScamCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/cv/analyze": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Cv Analyze",
        "operationId": "cv_analyze_api_v1_ai_cv_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CvAnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CvAnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/disaster-risk": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Disaster Risk",
        "operationId": "disaster_risk_api_v1_ai_disaster_risk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisasterRiskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisasterRiskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/decision-graph": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Decision Graph",
        "operationId": "decision_graph_api_v1_ai_decision_graph_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionGraphRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionGraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/feed-items": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Feed Item",
        "operationId": "create_feed_item_api_v1_admin_feed_items_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/feed-items/import-csv": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Import Feed Csv",
        "description": "Import verified opportunities from the documented CSV template.\n\nIntended for official scholarship/job records that do not expose an API.\nThe endpoint validates each row through ``FeedItemCreate`` and caps upload\nsize/row count to limit abuse.",
        "operationId": "import_feed_csv_api_v1_admin_feed_items_import_csv_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_feed_csv_api_v1_admin_feed_items_import_csv_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/refresh": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Refresh",
        "operationId": "refresh_api_v1_admin_refresh_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CvAnalyzeRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 20,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CvAnalyzeRequest"
      },
      "CvAnalyzeResponse": {
        "properties": {
          "extracted_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extracted Skills"
          },
          "detected_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Email"
          },
          "detected_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Phone"
          },
          "experience_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Years"
          },
          "recommended_roles": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  }
                ]
              },
              "type": "object"
            },
            "type": "array",
            "title": "Recommended Roles"
          }
        },
        "type": "object",
        "required": [
          "extracted_skills",
          "detected_email",
          "detected_phone",
          "experience_years",
          "recommended_roles"
        ],
        "title": "CvAnalyzeResponse"
      },
      "DecisionGraphRequest": {
        "properties": {
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Skills"
          },
          "preferred_categories": {
            "items": {
              "type": "string",
              "enum": [
                "job",
                "scholarship",
                "disaster",
                "weather",
                "service",
                "safety",
                "learning"
              ]
            },
            "type": "array",
            "maxItems": 10,
            "title": "Preferred Categories"
          },
          "country": {
            "type": "string",
            "maxLength": 100,
            "title": "Country",
            "default": ""
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "title": "City",
            "default": ""
          },
          "study_level": {
            "type": "string",
            "maxLength": 80,
            "title": "Study Level",
            "default": ""
          },
          "field_of_study": {
            "type": "string",
            "maxLength": 120,
            "title": "Field Of Study",
            "default": ""
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Interests"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 20
          },
          "max_items": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Items",
            "default": 15
          }
        },
        "type": "object",
        "title": "DecisionGraphRequest"
      },
      "DecisionGraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "top_items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationOut"
            },
            "type": "array",
            "title": "Top Items"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges",
          "top_items",
          "explanation"
        ],
        "title": "DecisionGraphResponse"
      },
      "DisasterRiskRequest": {
        "properties": {
          "rainfall_mm": {
            "type": "number",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Rainfall Mm"
          },
          "river_level_m": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "River Level M"
          },
          "soil_moisture": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Soil Moisture"
          },
          "slope_degrees": {
            "type": "number",
            "maximum": 90.0,
            "minimum": 0.0,
            "title": "Slope Degrees"
          },
          "temperature_c": {
            "type": "number",
            "maximum": 70.0,
            "minimum": -80.0,
            "title": "Temperature C"
          },
          "wind_kph": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Wind Kph"
          }
        },
        "type": "object",
        "required": [
          "rainfall_mm",
          "river_level_m",
          "soil_moisture",
          "slope_degrees",
          "temperature_c",
          "wind_kph"
        ],
        "title": "DisasterRiskRequest"
      },
      "DisasterRiskResponse": {
        "properties": {
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Risk Level"
          },
          "drivers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Drivers"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "type": "object",
        "required": [
          "risk_score",
          "risk_level",
          "drivers",
          "model_version"
        ],
        "title": "DisasterRiskResponse"
      },
      "FeedItemCreate": {
        "properties": {
          "external_id": {
            "type": "string",
            "maxLength": 180,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "External Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "job",
              "scholarship",
              "disaster",
              "weather",
              "service",
              "safety",
              "learning"
            ],
            "title": "Category"
          },
          "title": {
            "type": "string",
            "maxLength": 250,
            "minLength": 5,
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 10,
            "title": "Summary"
          },
          "source_name": {
            "type": "string",
            "maxLength": 180,
            "minLength": 2,
            "title": "Source Name"
          },
          "source_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Source Url"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "location": {
            "type": "string",
            "maxLength": 180,
            "title": "Location",
            "default": "Global"
          },
          "country_code": {
            "type": "string",
            "maxLength": 8,
            "title": "Country Code",
            "default": ""
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Tags"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Severity",
            "default": "low"
          },
          "verification_status": {
            "type": "string",
            "enum": [
              "unverified",
              "verified",
              "demo",
              "community"
            ],
            "title": "Verification Status",
            "default": "unverified"
          },
          "source_reliability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Source Reliability",
            "default": 0.5
          },
          "funding_type": {
            "type": "string",
            "maxLength": 80,
            "title": "Funding Type",
            "default": ""
          },
          "study_level": {
            "type": "string",
            "maxLength": 80,
            "title": "Study Level",
            "default": ""
          },
          "employment_type": {
            "type": "string",
            "maxLength": 80,
            "title": "Employment Type",
            "default": ""
          },
          "salary_text": {
            "type": "string",
            "maxLength": 120,
            "title": "Salary Text",
            "default": ""
          },
          "eligibility": {
            "type": "string",
            "maxLength": 2000,
            "title": "Eligibility",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "external_id",
          "category",
          "title",
          "summary",
          "source_name",
          "source_url",
          "published_at"
        ],
        "title": "FeedItemCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password"
          },
          "display_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Display Name"
          },
          "country": {
            "type": "string",
            "maxLength": 100,
            "title": "Country",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "display_name"
        ],
        "title": "RegisterRequest"
      },
...
      "ScamCheckRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 3,
            "title": "Text"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ScamCheckRequest"
      },
      "ScamCheckResponse": {
        "properties": {
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Level"
          },
          "label": {
            "type": "string",
            "enum": [
              "likely_legitimate",
              "suspicious",
              "likely_scam"
            ],
            "title": "Label"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          },
          "safe_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Safe Actions"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "type": "object",
        "required": [
          "risk_score",
          "risk_level",
          "label",
          "evidence",
          "safe_actions",
          "model_version"
        ],
        "title": "ScamCheckResponse"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}